refactor: splitted BookControllerAsText
"It seems that the reviewed classes of the controller package are
breaking the Single responsibility principle given that they
perform more than one task"
-- Quote from report "Search And Reporting"

Following the conclusion of that report, we've refactored the class
BookControllerAsText. This has helped with inversion of control and
reduced the cohesion of the mentionend class.
