+++ LibrarySystem/.gitignore
@@ -1,3 +1,4 @@
 obj
 bin
 out
+.idea
