+++ LibrarySystem/Program.cs
@@ -1,4 +1,6 @@
-﻿namespace LibrarySystem;
+using LibrarySystem.Controller.BookManagement;
+
+namespace LibrarySystem;
 
 class Program
 {
@@ -48,11 +50,18 @@ class Program
         EntityRendererAsConsolePages<Book> bookRendererAsPages = new EntityRendererAsConsolePages<Book>(bookRepository, bookFormatterFactory);
         EntityRendererAsConsolePages<Patron> patronRendererAsPages = new EntityRendererAsConsolePages<Patron>(patronRepository, patronFormatterFactory);
 
-        IExecutableHandler<string> bookController = new BookControllerAsText(bookRepository, bookCreator, bookUpdater, bookEliminator, bookFormatterFactory, messageRenderer, bookRendererAsPages);
         IExecutableHandler<string> patronController = new PatronControllerAsText(patronRepository, patronCreator, patronUpdater, patronEliminator, messageRenderer, patronFormatterFactory, patronRendererAsPages);
         IExecutableHandler<string> lenderController = new LoanControllerAsText(lender, loanRepository, patronRepository, bookRepository, messageRenderer, patronSelectorByConsole, bookSelectorByConsole);
         IExecutableHandler<string> fineController = new FineControllerAsText(debtManager, fineRepository, messageRenderer, fineFormatterFactory, fineSelectorByConsole);
         IExecutableHandler<string> reportController = new ReporterControllerAsText(reporter, statisticsGenerator, patronRepository, bookFormatterFactory, patronFormatterFactory, loanFormatterFactory, messageRenderer, patronSelectorByConsole, fineFormatterFactory);
+        IExecutableHandler<string> bookController = new BookControllerAsText(
+            messageRenderer,
+            new BookSearcher(bookRepository, bookFormatterFactory),
+            new BookRepositoryHandler(bookCreator, bookUpdater, bookEliminator),
+            bookRepository,
+            bookFormatterFactory,
+            bookRendererAsPages);
+        
 
         Dictionary<MenuView, IExecutableHandler<string>> controllers = new Dictionary<MenuView, IExecutableHandler<string>>
         {
